Backend integrations
previous backend had a lot of issues some of which i was unable to fix, so i wrote a new backend that is properly working. To test the project simply download my upload, ensure you have nodejs installed, migrate to the root of the folder and type 'npm install' (without the quotes of course), wait for installation to complete then type' npm run devStart' or 'nodemon app.js ' wait for the connection to database sucessful then migrate your browser to localhost:5000.